Remove internal/package.scala
