+++ core/src/main/scala/org/scalafmt/internal/FormatToken.scala
@@ -1,6 +1,7 @@
 package org.scalafmt.internal
 
 import org.scalafmt.internal.ScalaFmtLogger._
+import org.scalafmt.internal.TokenOps._
 
 import scala.collection.mutable
 import scala.meta.tokens.Token
