+++ core/src/main/scala/org/scalafmt/internal/Split.scala
@@ -1,5 +1,6 @@
 package org.scalafmt.internal
 
+import org.scalafmt.internal.Policy.NoPolicy
 import scala.meta.tokens.Token
 
 case class OptimalToken(token: Token, killOnFail: Boolean = false)
