+++ core/src/main/scala/org/scalafmt/internal/TokenOps.scala
@@ -16,6 +16,34 @@ object TokenOps extends TokenOps
   */
 trait TokenOps {
 
+  /**
+    * For convenience when experimenting with different hashing strategies.
+    */
+  type TokenHash = Long
+
+  /**
+    * Custom hash code for token.
+    *
+    * The default hashCode is slow because it prevents conflicts between
+    * tokens from different source files. We only care about getting a unique
+    * identifier for the token inside this source file.
+    *
+    * The hash code works like this this:
+    * Top 8 bits go to privateTag, a unique identifier for the tokens class.
+    * Next 28 bits go to the tokens **start** offset byte.
+    * Final 28 bits go to the tokens **end** offset byte.
+    *
+    * The only chance for collision is if two empty length tokens with the same
+    * type lie next to each other. @xeno-by said this should not happen.
+    */
+  @inline
+  def hash(token: Token): TokenHash = {
+    val longHash: Long =
+      (token.privateTag.toLong << (62 - 8)) |
+      (token.start.toLong << (62 - (8 + 28))) | token.end
+    longHash
+  }
+
   def shouldGet2xNewlines(tok: FormatToken): Boolean = {
     !isDocstring(tok.left) && {
       val newlines = newlinesBetween(tok.between)
